fix: ТЗ15. Исправлены замечания по ревью.

+++ src/main/java/ru/practicum/shareit/item/model/ItemAnswer.java
@@ -7,7 +7,7 @@ public interface ItemAnswer {
 
     String getDescription();
 
-    Integer getRequestId();
+    Long getRequestId();
 
     Boolean getAvailable();
 }
